requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Templates and static files
jinja2==3.1.3
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25

# Authentication
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Audio processing
pydub==0.25.1
soundfile==0.12.1

# AI/ML
openai==1.10.0
chromadb==0.4.22
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
